V0.2.8.1 - Added Backend Sources
Added:
  - ClassSelection model
  - Swapped BatchSelectionRecyclerAdapter & BatchRegisterRecyclerAdapter
  - FirebaseAuthentication to handle authentication
  - BackendTestingCode to test firebase

Some minor fixed & improvements

TIP: Commit not fully prepared still under progress

Further Version Improvements
  - Add Comments
  - Rename Variables Correctly
  - Refactor all Files with Code Formatting
  - Recheck the Functions to scope for Efficiency Improvement & Complexity Reduction
