+++ app/src/main/java/com/shubhamdeshmukh/attendencemanagementsystem/backend/database_entities/Class.java
@@ -87,4 +87,8 @@ public class Class {
     public void setBatchList(ArrayList<Batch> batchList) {
         this.batchList = batchList;
     }
+
+    public void setName(String name) {
+        this.name = name;
+    }
 }
