+++ app/src/main/java/com/shubhamdeshmukh/attendencemanagementsystem/frontend/teacherregistration/RegisterClassInfoAndBatchesActivity.java
@@ -23,7 +23,6 @@ import com.shubhamdeshmukh.attendencemanagementsystem.backend.FirebaseDBConnecti
 import com.shubhamdeshmukh.attendencemanagementsystem.backend.database_entities.Batch;
 import com.shubhamdeshmukh.attendencemanagementsystem.backend.database_entities.Class;
 import com.shubhamdeshmukh.attendencemanagementsystem.backend.database_entities.Data;
-import com.shubhamdeshmukh.attendencemanagementsystem.backend.models.BatchSelection;
 import com.shubhamdeshmukh.attendencemanagementsystem.frontend.MainActivity;
 
 import java.util.ArrayList;
@@ -33,7 +32,7 @@ public class RegisterClassInfoAndBatchesActivity extends AppCompatActivity {
     ArrayList<Batch> batchArrayList;
     int selectedClassIndex;
 
-    BatchSelectionRecyclerAdapter currentAdapter;
+    BatchRegisterRecyclerAdapter currentAdapter;
 
 
     @Override
@@ -47,21 +46,23 @@ public class RegisterClassInfoAndBatchesActivity extends AppCompatActivity {
             return insets;
         });
 
-        FloatingActionButton fab = findViewById(R.id.addbatch);
+        FloatingActionButton addBatchFloatingButton = findViewById(R.id.addbatch);
 
         Intent intent = getIntent();
-        FirebaseDBConnection firebaseDBConnection = new FirebaseDBConnection(MainActivity.database,MainActivity.mAuth);
         selectedClassIndex = intent.getIntExtra("classFetchedDataIndex", -1);
+        EditText className = findViewById(R.id.class_name);
 
         if (selectedClassIndex == -1) {
 
-            batchArrayList  = firebaseDBConnection.getRegistrationData().batches;
+            Data registrationData = MainActivity.dbConnection.getRegistrationData();
+            registrationData.classes.add(new Class());
+            selectedClassIndex = registrationData.classes.size() - 1;
+            batchArrayList  = new ArrayList<>();
 
         }
         else
         {
-            Class _class = firebaseDBConnection.getRegistrationData().classes.get(selectedClassIndex);
-            EditText className = findViewById(R.id.class_name);
+            Class _class = MainActivity.dbConnection.getRegistrationData().classes.get(selectedClassIndex);
             className.setText(_class.getName());
             batchArrayList = _class.getBatchList();
         }
@@ -71,11 +72,11 @@ public class RegisterClassInfoAndBatchesActivity extends AppCompatActivity {
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
 
-        currentAdapter = new BatchSelectionRecyclerAdapter(batchArrayList, this);
+        currentAdapter = new BatchRegisterRecyclerAdapter(batchArrayList, this);
         recyclerView.setAdapter(currentAdapter);
 
 
-        fab.setOnClickListener(new View.OnClickListener() {
+        addBatchFloatingButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
 
@@ -83,21 +84,20 @@ public class RegisterClassInfoAndBatchesActivity extends AppCompatActivity {
             }
         });
 
-        Button saveButton = findViewById(R.id.saveButton);
-        saveButton.setOnClickListener(new View.OnClickListener() {
+        Button submitDataButton = findViewById(R.id.submitDataButton);
+        submitDataButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Data registrationData = firebaseDBConnection.getRegistrationData();
-                ArrayList<Batch> newBatchList = currentAdapter.getBatchArrayList();;
-                ArrayList<Batch> registeredBatchList = new ArrayList<>();
+                Data registrationData = MainActivity.dbConnection.getRegistrationData();
 
-
-                registrationData.classes.get(selectedClassIndex).setBatchList(newBatchList);
-                registrationData.batches = registeredBatchList;
-                firebaseDBConnection.setRegistrationData(registrationData);
-                firebaseDBConnection.completeRegistration();
+                registrationData.classes.get(selectedClassIndex).setName(className.getText().toString());
+                registrationData.classes.get(selectedClassIndex).setBatchList(batchArrayList);
+                registrationData.batches.addAll(batchArrayList);
+                MainActivity.dbConnection.setRegistrationData(registrationData);
+                MainActivity.dbConnection.completeRegistration();
                 FirebaseDBConnection.updateDatabase();
-                Log.d(MainActivity.TAG, "onClick: COMPLETED REGISTRATION!");
+                Log.d(MainActivity.TAG, "onClick: COMPLETED CLASS REGISTRATION!");
+                finish();
             }
         });
     }
@@ -106,7 +106,7 @@ public class RegisterClassInfoAndBatchesActivity extends AppCompatActivity {
     {
         RecyclerView recyclerView = findViewById(R.id.register_batch_recycler_view);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        recyclerView.setAdapter(new BatchSelectionRecyclerAdapter(batchArrayList, this));
+        recyclerView.setAdapter(new BatchRegisterRecyclerAdapter(batchArrayList, this));
     }
 
     public void showBatchInfoDialog(int selectedBatch) {
@@ -149,37 +149,33 @@ public class RegisterClassInfoAndBatchesActivity extends AppCompatActivity {
                 String startingStudentNo = editTextStartingStudentNo.getText().toString();
                 String endingStudentNo = editTextEndingStudentNo.getText().toString();
 
-                FirebaseDBConnection firebaseDBConnection = new FirebaseDBConnection(MainActivity.database,MainActivity.mAuth);
-
                 // Update the batch & class info in registration Data in FirebaseDBConnection...
-                Data registrationData = firebaseDBConnection.getRegistrationData();
+                Data registrationData = MainActivity.dbConnection.getRegistrationData();
 
                 if (selectedClassIndex != -1)
                 {
                     if (selectedBatch != -1)
                     {
                         // Both class & batch are selected
-                        registrationData.classes.get(selectedClassIndex).getBatchList().get(selectedBatch).name = batchName;
-                        registrationData.classes.get(selectedClassIndex).getBatchList().get(selectedBatch).startId = startingStudentNo;
-                        registrationData.classes.get(selectedClassIndex).getBatchList().get(selectedBatch).endId = endingStudentNo;
+                        batchArrayList.get(selectedBatch).name = batchName;
+                        batchArrayList.get(selectedBatch).startId = startingStudentNo;
+                        batchArrayList.get(selectedBatch).endId = endingStudentNo;
                     }
                     else
                     {
                         // Only Class is selected
                         Batch newBatch = new Batch(batchName, startingStudentNo, endingStudentNo);
-                        registrationData.classes.get(selectedClassIndex).addBatch(newBatch);
+                        batchArrayList.add(newBatch);
                     }
                 }
                 else
                 {
                     // None is selected
-                    Class newClass = new Class();
-                    newClass.addBatch(new Batch(batchName, startingStudentNo, endingStudentNo));
-                    registrationData.classes.add(newClass);
-                    selectedClassIndex = registrationData.classes.size() - 1;
+//                    Class newClass = new Class();
+//                    newClass.addBatch(new Batch(batchName, startingStudentNo, endingStudentNo));
+//                    registrationData.classes.add(newClass);
+//                    selectedClassIndex = registrationData.classes.size() - 1;
                 }
-                firebaseDBConnection.setRegistrationData(registrationData);
-                ArrayList<Batch> batchArrayList = registrationData.classes.get(selectedClassIndex).getBatchList();
                 updateRecycler();
                 // Perform action with the input data
                 // For example, you can validate inputs or submit them to a server
