+++ app/src/main/res/layout/activity_register_class_info_and_batches.xml
@@ -60,7 +60,7 @@
             app:layout_constraintEnd_toEndOf="parent" />
 
         <Button
-            android:id="@+id/saveButton"
+            android:id="@+id/submitDataButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="15dp"
